pieni fiksaus

+++ src/main/java/projekti/demo/web/TapahtumaRestController.java
@@ -3,16 +3,20 @@ package projekti.demo.web;
 //import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
 
 import projekti.demo.model.Tapahtuma;
 import projekti.demo.model.TapahtumaRepository;
 
 @RestController
+@RequestMapping("/api/tapahtumat")
 public class TapahtumaRestController {
 
     @Autowired
@@ -20,9 +24,14 @@ public class TapahtumaRestController {
 
     // poista tapahtuma
     @DeleteMapping("/api/tapahtumat/{id}")
-    public Iterable<Tapahtuma> deleteTapahtuma(@PathVariable Long id) {
-        tapahtumaRepository.deleteById(id);
-        return tapahtumaRepository.findAll();
+    public ResponseEntity<Void> deleteTapahtuma(@PathVariable Long id) {
+        if (tapahtumaRepository.existsById(id)) {
+            tapahtumaRepository.deleteById(id);
+               return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     }
 
     //muokkaa tapahtumaa
@@ -34,5 +43,6 @@ public class TapahtumaRestController {
     }
 
     //näin jossain käytettävän public ResponseEntity<Tapahtuma> - onko tietoa olisiko tämä parempi?
+    //ResponseEntity ainakin mahdollistaa paluuarvojen antamisen varmaankin paremmin
 
 }
